requirements.txt
attrs==21.2.0
backcall==0.2.0
bleach==4.0.0
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cycler==0.10.0
Cython==0.29.24
decorator==4.4.2
defusedxml==0.7.1
einops==0.3.0
entrypoints==0.3
filelock==3.0.12
grpcio==1.39.0
idna==3.2
imutils==0.5.4
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
pycparser==2.20
Pygments
pymssql==2.2.2
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz
PyYAML==5.4.1
pyzmq==19.0.2
qtconsole==5.1.1
QtPy==1.9.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
terminado==0.11.0
testpath==0.5.0
tornado==6.1
tqdm==4.62.0
traitlets
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
uvicorn==0.17.6
fastapi==0.78.0
aiofiles
python-multipart==0.0.5
bcrypt==3.2.2
passlib==1.7.4
python-jose==3.0.0
celery==5.2.6
kombu
redis
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
locust
python-dotenv
gunicorn
scikit-learn==1.1.0
PyMySQL==1.0.2
plotly==5.8.0
fastapi-utils==0.2.1
apscheduler
scipy==1.8.0
openpyxl
pmdarima